In getCB, don't use data if the get did not return ECA_NORMAL
This seems to prevent us trying to pass the bad data received
back to the client, which can lead to a memcpy of bad data
and a segmentation fault in the gateway.

Prompted by looking at #20 and changes made in c6dc15.
In which: status = ECA_TOLARGE, "The requested
data transfer is greater than available memory or
EPICS_CA_MAX_ARRAY_BYTES" which is seen when
EPICS_CA_MAX_ARRAY_BYTES is set too small on the IOC hsoting the
requested PV.