Support an EPICS base and PCAS with IP access groups.
If EPICS base and PCAS have IP access groups implemented, the APIs of both need
the client's IP address as parameter. This version of the channel access server
provides this parameter.

Notes for the dependency from EPICS base and PCAS:

The extra parameter is only provided when EPICS_HAS_AS_IPAG is defined, which
is a C macro that is set in EPICS base. If this macro is not defined, the old
API is used. By this, this version of the channel access gateway can also be
compiled with an older version of EPICS base and PCAS.